XCore: update scotch to v7
git-svn-id: https://elsa-svn.onera.fr/CASSIOPEE/Trunk/Cassiopee/Apps/Modules@4826 2bbd7fe0-4929-0410-bb6b-e55522daae51

+++ XCore/XCore/scotch/dgraph_ghst.h
@@ -1,4 +1,4 @@
-/* Copyright 2007 ENSEIRB, INRIA & CNRS
+/* Copyright 2007,2021 IPB, Universite de Bordeaux, INRIA & CNRS
 **
 ** This file is part of the Scotch software package for static mapping,
 ** graph partitioning and sparse matrix ordering.
@@ -8,13 +8,13 @@
 ** use, modify and/or redistribute the software under the terms of the
 ** CeCILL-C license as circulated by CEA, CNRS and INRIA at the following
 ** URL: "http://www.cecill.info".
-** 
+**
 ** As a counterpart to the access to the source code and rights to copy,
 ** modify and redistribute granted by the license, users are provided
 ** only with a limited warranty and the software's author, the holder of
 ** the economic rights, and the successive licensors have only limited
 ** liability.
-** 
+**
 ** In this respect, the user's attention is drawn to the risks associated
 ** with loading, using, modifying and/or developing or reproducing the
 ** software by the user in light of its specific status of free software,
@@ -25,32 +25,35 @@
 ** their requirements in conditions enabling the security of their
 ** systems and/or data to be ensured and, more generally, to use and
 ** operate it in the same conditions as regards security.
-** 
+**
 ** The fact that you are presently reading this means that you have had
 ** knowledge of the CeCILL-C license and that you accept its terms.
 */
-/**********************************************************/
-/*                                                        */
-/*   NAME       : dgraph_ghst.h                           */
-/*                                                        */
-/*   AUTHOR     : Francois PELLEGRINI                     */
-/*                Francois CHATENET (P0.0)                */
-/*                Sebastien FOUCAULT (P0.0)               */
-/*                Nicolas GICQUEL (P0.1)                  */
-/*                Jerome LACOSTE (P0.1)                   */
-/*                                                        */
-/*   FUNCTION   : Part of a parallel static mapper.       */
-/*                These lines are the data declarations   */
-/*                for the halo building routine.          */
-/*                                                        */
-/*                # Version P0.0 : from : 01 apr 1997     */
-/*                                 to     20 jun 1997     */
-/*                # Version P0.1 : from : 12 apr 1998     */
-/*                                 to     20 jun 1998     */
-/*                # Version 5.0  : from : 28 feb 2006     */
-/*                                 to     31 dec 2006     */
-/*                                                        */
-/**********************************************************/
+/************************************************************/
+/**                                                        **/
+/**   NAME       : dgraph_ghst.h                           **/
+/**                                                        **/
+/**   AUTHOR     : Francois PELLEGRINI                     **/
+/**                Francois CHATENET (P0.0)                **/
+/**                Sebastien FOUCAULT (P0.0)               **/
+/**                Nicolas GICQUEL (P0.1)                  **/
+/**                Jerome LACOSTE (P0.1)                   **/
+/**                                                        **/
+/**   FUNCTION   : These lines are the data declarations   **/
+/**                for the halo building routine.          **/
+/**                                                        **/
+/**                # Version P0.0 : from : 01 apr 1997     **/
+/**                                 to   : 20 jun 1997     **/
+/**                # Version P0.1 : from : 12 apr 1998     **/
+/**                                 to   : 20 jun 1998     **/
+/**                # Version 5.0  : from : 28 feb 2006     **/
+/**                                 to   : 31 dec 2006     **/
+/**                # Version 6.1  : from : 04 dec 2021     **/
+/**                                 to   : 04 dec 2021     **/
+/**                # Version 7.0  : from : 22 mar 2023     **/
+/**                                 to   : 22 mar 2023     **/
+/**                                                        **/
+/************************************************************/
 
 /*
 ** The defines.
@@ -58,7 +61,7 @@
 
 /* procsidtab-related values. */
 
-#define DGRAPHGHSTSIDMAX            ((int) ((unsigned int) (1 << (sizeof (int) * 8 - 1)) - 2U)) /* Maximum leap value for procsidtab entries */
+#define DGRAPHGHSTSIDMAX            ((int) ((1U << (sizeof (unsigned int) * 8 - 1)) - 2U)) /* Maximum leap value for procsidtab entries */
 
 /*
 ** The type and structure definitions.
